+++ clients/admin-ui/cypress/e2e/action-center.cy.ts
@@ -2,6 +2,7 @@ import {
   stubActionCenter,
   stubPlus,
   stubSystemVendors,
+  stubTaxonomyEntities,
   stubVendorList,
 } from "cypress/support/stubs";
 
@@ -181,7 +182,9 @@ describe("Action center", () => {
       cy.getByTestId("row-0-col-actions").within(() => {
         cy.getByTestId("ignore-btn").click({ force: true });
       });
-      cy.wait("@ignoreMonitorResultUncategorizedSystem");
+      cy.wait("@ignoreMonitorResultSystem").then((interception) => {
+        expect(interception.request.url).to.contain("[undefined]");
+      });
       cy.getByTestId("success-alert").should(
         "contain",
         "108 uncategorized assets have been ignored and will not appear in future scans.",
@@ -207,8 +210,42 @@ describe("Action center", () => {
         "10 assets from Google Tag Manager have been ignored and will not appear in future scans.",
       );
     });
+    it("shouldn't allow bulk add when uncategorized system is selected", () => {
+      cy.getByTestId("row-0-col-select").find("label").click();
+      cy.getByTestId("selected-count").should("contain", "1 selected");
+      cy.getByTestId("bulk-actions-menu").click();
+      cy.getByTestId("bulk-add").should("be.disabled");
+    });
+    it("should bulk add results from categorized systems", () => {
+      cy.getByTestId("bulk-actions-menu").should("be.disabled");
+      cy.getByTestId("row-1-col-select").find("label").click();
+      cy.getByTestId("row-2-col-select").find("label").click();
+      cy.getByTestId("selected-count").should("contain", "2 selected");
+      cy.getByTestId("bulk-actions-menu").should("not.be.disabled");
+      cy.getByTestId("bulk-actions-menu").click();
+      cy.getByTestId("bulk-add").click();
+      cy.wait("@addMonitorResultSystem");
+      cy.getByTestId("success-alert").should(
+        "contain",
+        "16 assets have been added to the system inventory.",
+      );
+    });
+    it("should bulk ignore results from all systems", () => {
+      cy.getByTestId("row-0-col-select").find("label").click();
+      cy.getByTestId("row-1-col-select").find("label").click();
+      cy.getByTestId("row-2-col-select").find("label").click();
+      cy.getByTestId("selected-count").should("contain", "3 selected");
+      cy.getByTestId("bulk-actions-menu").should("not.be.disabled");
+      cy.getByTestId("bulk-actions-menu").click();
+      cy.getByTestId("bulk-ignore").click();
+      cy.wait("@ignoreMonitorResultSystem");
+      cy.getByTestId("success-alert").should(
+        "contain",
+        "124 assets have been ignored and will not appear in future scans.",
+      );
+    });
     it("should navigate to table view on row click", () => {
-      cy.getByTestId("row-1").click();
+      cy.getByTestId("row-1-col-system_name").click();
       cy.url().should(
         "contain",
         "system_key-8fe42cdb-af2e-4b9e-9b38-f75673180b88",
@@ -239,8 +276,7 @@ describe("Action center", () => {
       cy.getByTestId("column-name").should("exist");
       cy.getByTestId("column-resource_type").should("exist");
       cy.getByTestId("column-system").should("exist");
-      // TODO: [HJ-369] uncomment when data use column is implemented
-      // cy.getByTestId("column-data_use").should("exist");
+      cy.getByTestId("column-data_use").should("exist");
       cy.getByTestId("column-locations").should("exist");
       cy.getByTestId("column-domain").should("exist");
       // TODO: [HJ-344] uncomment when Discovery column is implemented
@@ -261,7 +297,7 @@ describe("Action center", () => {
       });
       cy.wait("@getSystemsPaginated");
       cy.getByTestId("system-select").antSelect("Fidesctl System");
-      cy.wait("@setAssetSystem");
+      cy.wait("@patchAssets");
       cy.getByTestId("system-select").should("not.exist");
       cy.getByTestId("success-alert").should(
         "contain",
@@ -290,8 +326,7 @@ describe("Action center", () => {
       cy.getByTestId("column-name").should("exist");
       cy.getByTestId("column-resource_type").should("exist");
       cy.getByTestId("column-system").should("exist");
-      // TODO: [HJ-369] uncomment when data use column is implemented
-      // cy.getByTestId("column-data_use").should("exist");
+      cy.getByTestId("column-data_use").should("exist");
       cy.getByTestId("column-locations").should("exist");
       cy.getByTestId("column-domain").should("exist");
       // TODO: [HJ-344] uncomment when Discovery column is implemented
@@ -312,7 +347,7 @@ describe("Action center", () => {
       });
       cy.wait("@getSystemsPaginated");
       cy.getByTestId("system-select").antSelect("Fidesctl System");
-      cy.wait("@setAssetSystem");
+      cy.wait("@patchAssets");
       cy.getByTestId("system-select").should("not.exist");
       cy.getByTestId("success-alert").should(
         "contain",
@@ -333,7 +368,7 @@ describe("Action center", () => {
       });
       cy.wait("@getSystemsPaginated");
       cy.getByTestId("system-select").antSelect("Demo Marketing System");
-      cy.wait("@setAssetSystem");
+      cy.wait("@patchAssets");
       cy.getByTestId("success-alert").should("exist");
       cy.getByTestId("system-select").should("not.exist");
       cy.getByTestId("success-alert").should(
@@ -355,7 +390,7 @@ describe("Action center", () => {
       // adds new system
       cy.wait("@postSystemVendors");
       // assigns asset to new system
-      cy.wait("@setAssetSystem");
+      cy.wait("@patchAssets");
       cy.getByTestId("success-alert").should(
         "contain",
         'Test System has been added to your system inventory and the Browser Request "gtm.js" has been assigned to that system.',
@@ -430,6 +465,7 @@ describe("Action center", () => {
         "11 assets from Google Tag Manager have been added to the system inventory.",
       );
     });
+
     it("should bulk assign assets to a system", () => {
       cy.getByTestId("bulk-actions-menu").should("be.disabled");
       cy.getByTestId("row-0-col-select").find("label").click();
@@ -442,10 +478,29 @@ describe("Action center", () => {
       cy.getByTestId("add-modal-content").should("be.visible");
       cy.getByTestId("system-select").antSelect("Fidesctl System");
       cy.getByTestId("save-btn").click();
-      cy.wait("@setAssetSystem");
+      cy.wait("@patchAssets");
       cy.getByTestId("success-alert").should(
         "contain",
         "3 assets have been assigned to Fidesctl System.",
+      );
+    });
+
+    it("should bulk add data uses to assets", () => {
+      stubTaxonomyEntities();
+      cy.getByTestId("bulk-actions-menu").should("be.disabled");
+      cy.getByTestId("row-0-col-select").find("label").click();
+      cy.getByTestId("row-2-col-select").find("label").click();
+      cy.getByTestId("row-3-col-select").find("label").click();
+      cy.getByTestId("selected-count").should("contain", "3 selected");
+      cy.getByTestId("bulk-actions-menu").should("not.be.disabled");
+      cy.getByTestId("bulk-actions-menu").click();
+      cy.getByTestId("bulk-add-data-use").click();
+      cy.getByTestId("taxonomy-select").antSelect("essential");
+      cy.getByTestId("save-btn").click({ force: true });
+      cy.wait("@patchAssets");
+      cy.getByTestId("success-alert").should(
+        "contain",
+        "Consent categories added to 3 assets from Google Tag Manager.",
       );
     });
   });